Fix flow and improve tests.
