+++ src/main/java/solutions/TokenIssueFlowInitiator.java
@@ -1,6 +1,7 @@
-package bootcamp;
+package solutions;
 
 import co.paralleluniverse.fibers.Suspendable;
+import net.corda.core.contracts.CommandData;
 import net.corda.core.flows.*;
 import net.corda.core.identity.Party;
 import net.corda.core.transactions.SignedTransaction;
@@ -30,39 +31,28 @@ public class TokenIssueFlowInitiator extends FlowLogic<SignedTransaction> {
     @Suspendable
     @Override
     public SignedTransaction call() throws FlowException {
-        // We choose our transaction's notary (the notary prevents double-spends).
         Party notary = getServiceHub().getNetworkMapCache().getNotaryIdentities().get(0);
-        // We get a reference to our own identity.
+
         Party issuer = getOurIdentity();
 
-        /* ============================================================================
-         *         TODO 1 - Create our TokenState to represent on-ledger tokens!
-         * ===========================================================================*/
-        // We create our new TokenState.
-        TokenState tokenState = null;
+        TokenState tokenState = new TokenState(issuer, owner, amount);
 
+        TransactionBuilder transactionBuilder = new TransactionBuilder(notary);
 
-        /* ============================================================================
-         *      TODO 3 - Build our token issuance transaction to update the ledger!
-         * ===========================================================================*/
-        // We build our transaction.
-        TransactionBuilder transactionBuilder = null;
+        CommandData commandData = new TokenContract.Commands.Issue();
 
-        /* ============================================================================
-         *          TODO 2 - Write our TokenContract to control token issuance!
-         * ===========================================================================*/
-        // We check our transaction is valid based on its contracts.
+        transactionBuilder.addCommand(commandData, issuer.getOwningKey(), owner.getOwningKey());
+
+        transactionBuilder.addOutputState(tokenState, TokenContract.ID);
+
         transactionBuilder.verify(getServiceHub());
 
         FlowSession session = initiateFlow(owner);
 
-        // We sign the transaction with our private key, making it immutable.
         SignedTransaction signedTransaction = getServiceHub().signInitialTransaction(transactionBuilder);
 
-        // The counterparty signs the transaction
         SignedTransaction fullySignedTransaction = subFlow(new CollectSignaturesFlow(signedTransaction, singletonList(session)));
 
-        // We get the transaction notarised and recorded automatically by the platform.
         return subFlow(new FinalityFlow(fullySignedTransaction, singletonList(session)));
     }
 }
